Move Nullhandlig to its own class
