+++ pinot-segment-local/src/main/java/org/apache/pinot/segment/local/segment/index/nullvalue/NullValueIndexType.java
@@ -45,6 +45,7 @@ import org.apache.pinot.segment.spi.store.SegmentDirectory;
 import org.apache.pinot.spi.config.table.IndexConfig;
 import org.apache.pinot.spi.config.table.TableConfig;
 import org.apache.pinot.spi.data.FieldSpec;
+import org.apache.pinot.spi.data.NullHandling;
 import org.apache.pinot.spi.data.Schema;
 
 
@@ -84,7 +85,7 @@ public class NullValueIndexType extends AbstractIndexType<IndexConfig, NullValue
       Collection<FieldSpec> allFieldSpecs = schema.getAllFieldSpecs();
       Map<String, IndexConfig> configMap = Maps.newHashMapWithExpectedSize(allFieldSpecs.size());
 
-      Schema.NullHandling nullHandling = schema.getOptions().getNullHandling();
+      NullHandling nullHandling = schema.getOptions().getNullHandling();
 
       for (FieldSpec fieldSpec : allFieldSpecs) {
         IndexConfig indexConfig;
